k2a-django-base-setup/webapp/requirements.txt
Django==1.9.2
django-phonenumber-field==1.1.0
MySQL-python==1.2.5
mysqlclient==1.3.9
phonenumberslite==7.7.0
pytz==2016.6.1
requests==2.11.1
